Rename m_Val to m_SpecificVal; Allow m_Val to capture a given value if the pattern match. Add placeholders for AffineLoad/StoreOp.
We rename m_Val to m_SpecificVal similar to PatternMatch.h in LLVM. We add
another matcher m_Val which allows binding a Value to a variable if the pattern
is matched. We introduce placeholders.  The idea behind placeholders is to
allow the inspection of the access patterns in AffineStore/LoadOp. For now, we
limit ourself to simple affine function (i.e., coefficient * induction +
constant). We also provide another overload for bindDims which allow setting
the dimension at position pos.